Support usage of filepaths as -r and -u arguments.
Add `tsdb.name_from_path` so that if the user provides a filepath
as the -r or -u option, we can load their database from that
path and also figure out the subreddit / username to fetch the
subreddit / user object properly, without having to add another
cmd arg for specifying a nonstandard path. New constructor argument
fix_name to perform this automatically and return a tuple of
(database, fixed name).