Separate the generating of html from the writing it.
I was not happy with the fact that html_from_database actually
performed the file writes. Furthermore the other tool modules have
a titular function that performs the most main task but this one
didn't, it was hidden inside this function. So let's make
html_from_database just a generator and have a function
offline_reading that does what you came here to do.