Replace "> timesearch.py" with "python timesearch.py"
The bracket was meant to look like the command prompt, but for
linux users they would be more used to $ or #, and for people who
do not use the command line often they may be confused by it
altogether. This change should help make it a little more clear as
to what's going on.